+++ src/main/java/muse_kopis/muse/actor/domain/Actor.java
@@ -1,4 +1,4 @@
-package muse_kopis.muse.performance.domain.actor.domain;
+package muse_kopis.muse.actor.domain;
 
 import jakarta.persistence.Column;
 import jakarta.persistence.MappedSuperclass;
