+++ src/main/java/muse_kopis/muse/actor/domain/CastMemberRepository.java
@@ -1,4 +1,4 @@
-package muse_kopis.muse.performance.domain.actor.domain;
+package muse_kopis.muse.actor.domain;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
