+++ src/main/java/muse_kopis/muse/actor/domain/FavoriteActor.java
@@ -1,4 +1,4 @@
-package muse_kopis.muse.performance.domain.actor.domain;
+package muse_kopis.muse.actor.domain;
 
 import jakarta.persistence.*;
 import lombok.Getter;
