+++ src/main/java/muse_kopis/muse/photo/domain/dto/PhotoResponse.java
@@ -1,0 +1,6 @@
+package muse_kopis.muse.photo.domain.dto;
+
+public record PhotoResponse(
+        String url
+) {
+}
