--- src/main/java/muse_kopis/muse/ticketbook/domain/photo/domain/dto/PhotoResponse.java
@@ -1,6 +1,0 @@
-package muse_kopis.muse.ticketbook.domain.photo.domain.dto;
-
-public record PhotoResponse(
-        String url
-) {
-}
