+++ src/main/java/muse_kopis/muse/ticketbook/presentation/TicketBookController.java
@@ -8,7 +8,7 @@ import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import muse_kopis.muse.auth.Auth;
 import muse_kopis.muse.ticketbook.application.TicketBookService;
-import muse_kopis.muse.ticketbook.domain.photo.application.PhotoService;
+import muse_kopis.muse.photo.application.PhotoService;
 import muse_kopis.muse.ticketbook.infra.config.FrontURLConfig;
 import muse_kopis.muse.ticketbook.domain.dto.MonthDto;
 import muse_kopis.muse.ticketbook.domain.dto.ShareablePage;
